12.02.23: MeterUpdateView

+++ data/src/main/java/com/example/myhome/meter/dtos/MeterDto.kt
@@ -26,7 +26,7 @@ data class MeterGetDto(
     val issuedAt: String,
     val apartmentId: Int,
     val typeOfServiceId: Int,
-    val currentReading: Float?,
+    val currentReading: Double?,
     val typeOfServiceName: String,
     val unitName: String
 )
