+++ domain/src/main/java/com/example/myhome/meter/models/ReadingModel.kt
@@ -7,14 +7,14 @@ import java.util.Date
 abstract class ReadingModel(
     open val id: Int?,
     open val meterId: Int,
-    open val reading: Float,
+    open val reading: Double,
     open val readAt: Date
 )
 
 data class ReadingAddModel(
     override val id: Int?,
     override val meterId: Int,
-    override val reading: Float,
+    override val reading: Double,
     override val readAt: Date
 ) : ReadingModel(
     id = id,
@@ -25,7 +25,7 @@ data class ReadingAddModel(
 
 data class ReadingGetModel(
     val id: Int,
-    val reading: Float,
-    val consumption: Float,
+    val reading: Double,
+    val consumption: Double,
     val readAt: Date
 )
